+++ packages/firestore/src/api/database.ts
@@ -70,6 +70,7 @@ import {
   validateOptionalArgType,
   validateOptionalArrayElements,
   validateOptionNames,
+  validateStringEnum,
   valueDescription
 } from '../util/input_validation';
 import * as log from '../util/log';
@@ -109,6 +110,7 @@ import {
 const DEFAULT_HOST = 'firestore.googleapis.com';
 const DEFAULT_SSL = true;
 const DEFAULT_TIMESTAMPS_IN_SNAPSHOTS = true;
+const DEFAULT_FORCE_LONG_POLLING = false;
 
 /**
  * Constant used to indicate the LRU garbage collection should be disabled.
@@ -151,6 +153,8 @@ class FirestoreSettings {
 
   readonly cacheSizeBytes: number;
 
+  readonly forceLongPolling: boolean;
+
   // Can be a google-auth-library or gapi client.
   // tslint:disable-next-line:no-any
   credentials?: any;
@@ -177,7 +181,8 @@ class FirestoreSettings {
       'ssl',
       'credentials',
       'timestampsInSnapshots',
-      'cacheSizeBytes'
+      'cacheSizeBytes',
+      'experimentalForceLongPolling'
     ]);
 
     validateNamedOptionalType(
@@ -253,6 +258,17 @@ class FirestoreSettings {
         this.cacheSizeBytes = settings.cacheSizeBytes;
       }
     }
+
+    validateNamedOptionalType(
+      'settings',
+      'boolean',
+      'experimentalForceLongPolling',
+      settings.experimentalForceLongPolling
+    );
+    this.forceLongPolling =
+      settings.experimentalForceLongPolling === undefined
+        ? DEFAULT_FORCE_LONG_POLLING
+        : settings.experimentalForceLongPolling;
   }
 
   isEqual(other: FirestoreSettings): boolean {
@@ -261,7 +277,8 @@ class FirestoreSettings {
       this.ssl === other.ssl &&
       this.timestampsInSnapshots === other.timestampsInSnapshots &&
       this.credentials === other.credentials &&
-      this.cacheSizeBytes === other.cacheSizeBytes
+      this.cacheSizeBytes === other.cacheSizeBytes &&
+      this.forceLongPolling === other.forceLongPolling
     );
   }
 }
@@ -412,7 +429,8 @@ export class Firestore implements firestore.FirebaseFirestore, FirebaseService {
       this._config.databaseId,
       this._config.persistenceKey,
       this._config.settings.host,
-      this._config.settings.ssl
+      this._config.settings.ssl,
+      this._config.settings.forceLongPolling
     );
 
     const preConverter = (value: unknown) => {
@@ -1352,8 +1370,10 @@ export class Query implements firestore.Query {
     value: unknown
   ): firestore.Query {
     validateExactNumberOfArgs('Query.where', arguments, 3);
-    validateArgType('Query.where', 'non-empty string', 2, opStr);
     validateDefined('Query.where', 3, value);
+    // Enumerated from the WhereFilterOp type in index.d.ts.
+    const whereFilterOpEnums = ['<', '<=', '==', '>=', '>', 'array-contains'];
+    validateStringEnum('Query.where', whereFilterOpEnums, 2, opStr);
     let fieldValue;
     const fieldPath = fieldPathFromArgument('Query.where', field);
     const relationOp = RelationOp.fromString(opStr);