+++ packages/firestore/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Unreleased
+- [feature] Added an `experimentalForceLongPolling` setting that that can be
+  used to work around proxies that prevent the Firestore client from connecting
+  to the Firestore backend.
 
 # 1.1.1
 - [changed] Increased a connection timeout that could lead to large writes